requirements.txt
requests~=2.26.0
urwid~=2.1.2
python-dotenv~=0.19.2